requirement.txt
pandas
notebook
matplotlib
seaborn
numpy
PyYAML
psycopg2-binary
sqlalchemy
